Refactor : 에러 처리 분할
구입 금액, 당첨 번호, 보너스 번호 파일로 분할
